pyproject.toml
[project]
name = "tradinglab"
version = "0.1.0"
description = "Research & backtesting framework for trading strategies"
authors = [
    {name = "Alex Lebedev",email = "lebedevalex3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "ccxt (>=4.5.33,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
ruff = "^0.14.11"
pre-commit = "^4.5.1"
jupyterlab = "^4.5.2"
matplotlib = "^3.10.8"

[tool.ruff]
line-length = 200
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]